Changed data type for remainingPayments in model

+++ services/credit-service/credit-service/src/main/java/com/investmenttracker/creditservice/model/RepaymentProgress.java
@@ -14,6 +14,6 @@ public class RepaymentProgress {
     BigDecimal totalPaid;
     BigDecimal principalPaid;
     BigDecimal interestPaid;
-    BigDecimal remainingPayments;
+    Integer remainingPayments;
     BigDecimal remainingPercentage;
 }
